criando classe de erros personalizada

+++ src/storage/group/groupCreate.ts
@@ -1,5 +1,6 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 
+import { AppError } from "@utils/AppError";
 
 import { GROUP_COLLECTION } from "@storage/storageConfig";
 import { groupsGetAll } from "./groupsGetAll";
@@ -8,6 +9,12 @@ export const groupCreate = async (newGroup: string) => {
   try {
 
     const storedGroups = await groupsGetAll();
+
+    const groupAlreadyExists = storedGroups.includes(newGroup);
+
+    if(groupAlreadyExists) {
+      throw new AppError(`O grupo ${newGroup} já existe. Por favor, escolha outro nome.`);
+    }
 
     const storage = JSON.stringify([...storedGroups, newGroup]);
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
